+++ src/lib/app-provider/contexts/app.tsx
@@ -24,6 +24,8 @@ import {
 } from "env";
 import { changeFavicon } from "lib/utils";
 
+import { DEFAULT_CHAIN_CONFIG } from "./default";
+
 interface AppProviderProps {
   children: ReactNode;
 }
@@ -38,49 +40,6 @@ interface AppContextInterface {
   theme: ThemeConfig;
   setTheme: (newTheme: ThemeConfig) => void;
 }
-
-const DEFAULT_CHAIN_CONFIG: ChainConfig = {
-  tier: "lite",
-  chainId: "",
-  chain: "",
-  registryChainName: "",
-  prettyName: "",
-  lcd: "",
-  rpc: "",
-  wallets: [],
-  features: {
-    faucet: {
-      enabled: false,
-    },
-    wasm: {
-      enabled: false,
-    },
-    move: {
-      enabled: false,
-    },
-    pool: {
-      enabled: false,
-    },
-    gov: {
-      enabled: false,
-    },
-    nft: {
-      enabled: false,
-    },
-    publicProject: {
-      enabled: false,
-    },
-  },
-  gas: {
-    gasAdjustment: 1.0,
-    maxGasLimit: 0,
-  },
-  extra: {},
-  network_type: "testnet",
-  fees: {
-    fee_tokens: [],
-  },
-};
 
 const DEFAULT_STATES: AppContextInterface = {
   isHydrated: false,
