+++ src/lib/app-provider/hooks/index.ts
@@ -1,6 +1,8 @@
 export * from "./useAddress";
 export * from "./useDummyWallet";
 export * from "./useExampleAddresses";
+export * from "./useFabricateFee";
+export * from "./useGasPrice";
 export * from "./useInternalNavigate";
 export * from "./useLcdEndpoint";
 export * from "./useMediaQuery";
@@ -18,5 +20,3 @@ export * from "./useInitia";
 export * from "./useIsConnected";
 export * from "./useChainConfigs";
 export * from "./useAllowCustomNetworks";
-export * from "./useFabricateFee";
-export * from "./useGasPrice";
