+++ src/lib/types/index.ts
@@ -6,6 +6,7 @@ export * from "./addrs";
 export * from "./asset";
 export * from "./big";
 export * from "./block";
+export * from "./chainConfig";
 export * from "./code";
 export * from "./common";
 export * from "./contract";
@@ -26,4 +27,3 @@ export * from "./upload";
 export * from "./validator";
 export * from "./verification";
 export * from "./wallet";
-export * from "./chain-config";
