+++ src/lib/utils/index.ts
@@ -3,6 +3,7 @@ export * from "./array";
 export * from "./assetValue";
 export * from "./base64";
 export * from "./bech32";
+export * from "./chainConfig";
 export * from "./codeHash";
 export * from "./codePermission";
 export * from "./contractState";
@@ -41,4 +42,3 @@ export * from "./validator";
 export * from "./verification";
 export * from "./window";
 export * from "./zod";
-export * from "./chain-config";
